Creacion de interfaz Weapon
