+++ src/main/java/cl/uchile/dcc/TimerExample.java
@@ -2,9 +2,8 @@ package cl.uchile.dcc;
 
 import cl.uchile.dcc.finalreality.exceptions.InvalidStatValueException;
 import cl.uchile.dcc.finalreality.model.character.GameCharacter;
-import cl.uchile.dcc.finalreality.model.character.player.AbstractPlayerCharacter;
 import cl.uchile.dcc.finalreality.model.character.player.Thief;
-import cl.uchile.dcc.finalreality.model.weapon.Weapon;
+import cl.uchile.dcc.finalreality.model.weapon.DefaultWeapon;
 import cl.uchile.dcc.finalreality.model.weapon.WeaponType;
 import java.util.Random;
 import java.util.concurrent.BlockingQueue;
@@ -21,7 +20,7 @@ public class TimerExample {
     Random rng = new Random();
     for (int i = 0; i < 10; i++) {
       // Gives a random speed to each character to generate different waiting times
-      var weapon = new Weapon("", 0, rng.nextInt(50), WeaponType.KNIFE);
+      var weapon = new DefaultWeapon("", 0, rng.nextInt(50), WeaponType.KNIFE);
       var character = new Thief(Integer.toString(i), 10, 10, queue);
       character.equip(weapon);
       character.waitTurn();
