+++ src/main/java/cl/uchile/dcc/finalreality/model/character/player/PlayerCharacter.java
@@ -9,7 +9,7 @@ package cl.uchile.dcc.finalreality.model.character.player;
  */
 
 import cl.uchile.dcc.finalreality.model.character.GameCharacter;
-import cl.uchile.dcc.finalreality.model.weapon.Weapon;
+import cl.uchile.dcc.finalreality.model.weapon.DefaultWeapon;
 
 /**
  * A {@link GameCharacter} that can equip a weapon.
@@ -18,10 +18,10 @@ public interface PlayerCharacter extends GameCharacter {
   /**
    * Equips a weapon to the character.
    */
-  void equip(Weapon weapon);
+  void equip(DefaultWeapon defaultWeapon);
 
   /**
    * Return this character's equipped weapon.
    */
-  Weapon getEquippedWeapon();
+  DefaultWeapon getEquippedWeapon();
 }
